+++ src/appmixer/resend/core/ListAudiences/ListAudiences.js
@@ -1,18 +1,16 @@
-
 'use strict';
 
-const lib = require('../../../lib');
+const lib = require('../../lib');
 
-// Schema for a single domain item
+// Schema for a single audience item
 const schema = {
     id: { type: 'string', title: 'ID' },
     name: { type: 'string', title: 'Name' },
-    status: { type: 'string', title: 'Status' },
-    created_at: { type: 'string', title: 'Created At' },
-    region: { type: 'string', title: 'Region' }
+    created_at: { type: 'string', title: 'Created At' }
 };
 
 module.exports = {
+
     async receive(context) {
         const { outputType = 'array' } = context.messages.in.content || {};
 
@@ -22,26 +20,20 @@ module.exports = {
                 context,
                 outputType,
                 schema,
-                { label: 'Domains', value: 'result' }
+                { label: 'Audiences', value: 'result' }
             );
         }
 
         // Make the API request
         const response = await context.httpRequest({
             method: 'GET',
-            url: 'https://api.resend.com/domains',
+            url: 'https://api.resend.com/audiences',
             headers: {
                 'Authorization': `Bearer ${context.auth.apiKey}`
             }
         });
 
-        // Accepts both { data: [...] } and { data: { data: [...] } }
-        let items = [];
-        if (Array.isArray(response.data)) {
-            items = response.data;
-        } else if (response.data && Array.isArray(response.data.data)) {
-            items = response.data.data;
-        }
+        const items = response.data && Array.isArray(response.data) ? response.data : [];
 
         if (items.length === 0) {
             return context.sendJson({}, 'notFound');
@@ -53,4 +45,5 @@ module.exports = {
             outputType
         });
     }
+
 };
