+++ src/appmixer/resend/core/ListContacts/ListContacts.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const lib = require('../../../lib');
+const lib = require('../../lib');
 
 // Schema for a single contact item
 const schema = {
